requirements.txt
pandas
GitPython>=3.1.42
clinical_etl@git+https://github.com/CanDIG/clinical_ETL_code.git@v3.1.0
